ch-14: use sized numeric types

+++ src/debug.c
@@ -3,7 +3,7 @@
 #include "debug.h"
 #include "value.h"
 
-static int constant_instruction(const char* name, Chunk* chunk, int offset) {
+static int constant_instruction(const char* name, Chunk* chunk, size_t offset) {
     uint8_t idx = chunk->code[offset + 1];
     printf("%-16s %4d '", name, idx);
     print_value(chunk->constants.values[idx]);
@@ -11,17 +11,17 @@ static int constant_instruction(const char* name, Chunk* chunk, int offset) {
     return offset + 2;
 }
 
-static int simple_instruction(const char* name, int offset) {
+static int simple_instruction(const char* name, size_t offset) {
     printf("%s\n", name);
     return offset + 1;
 }
 
-int disassemble_instruction(Chunk* chunk, int offset) {
-    printf("%04d ", offset);
+int disassemble_instruction(Chunk* chunk, size_t offset) {
+    printf("%04zu ", offset);
     if (offset > 0 && chunk->lines[offset] == chunk->lines[offset - 1]) {
         printf("   | ");
     } else {
-        printf("%4d ", chunk->lines[offset]);
+        printf("%4zu ", chunk->lines[offset]);
     }
 
     uint8_t instruction = chunk->code[offset];
@@ -39,7 +39,7 @@ int disassemble_instruction(Chunk* chunk, int offset) {
 void disassemble_chunk(Chunk* chunk, const char* name) {
     printf("== %s ==\n", name);
 
-    int offset = 0;
+    size_t offset = 0;
     while (offset < chunk->count) {
         offset = disassemble_instruction(chunk, offset);
     }
